backend/functions/shared/requirements.txt
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
firebase-admin==6.8.0
google-auth==2.40.1